// client/src/modules/product/types/product.ts
export interface CartItem {
  id: string; // unique item key: productId-colorId-size
  productId: string;
  productName: string;
  brand: string;
  colorName: string;
  colorHex: string;
  size: string;
  price: number;
  originalPrice: number;
  image: string;
  sku: string;
  quantity: number;
}